finall

+++ src/main/java/com/kel5/ecommerce/entity/Product.java
@@ -24,7 +24,7 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private float price;
+    private long price;
     private int stock;
     private float weight;
